Function to create all of the plots and tables for the annual Canning River report.

canning_reportR takes an export filepath, data, and surface plot colour and produces panel plots and tables for all metrics.

Usage

canning_reportR(
  inpath,
  outpath,
  surface = "blue",
  bottom = "red",
  chloro = "darkgreen"
)

Arguments

inpath

character filepath to premade annual report data canning_WIN_report_data.

outpath

character filepath to desired export location.

surface

colour for surface plots. Can be named colour or hex format. Defaults to "blue".

bottom

colour for bottom plots. Can be named colour or hex format. Defaults to "red".

chloro

colour for integrated plots. Can be named colour or hex format. Defaults to "darkgreen".

Value

panel plots for all metrics and a csvs of metrics for inclusion to tables.

Details

This is a wrapper function that runs all of the individual functions to produce all of the plots and tables for the Canning River. Outputs will be exported to two folders created at the outpath location. `c_panels/` for plots and `c_tables/` for data tables.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/rivRmon/index.html the rivRmon website

Examples

if (FALSE) { # \dontrun{
canning_reportR(inpath, outpath, surface = "blue", bottom = "red",
chloro = "darkgreen")} # }