Reads in external data, combines with internal datasets and produces Canning River surfer plots

canning_surfR takes a file path to Canning River sonde output and creates a four panel (single column) surfer plot of salinity, dissolved oxygen, chlorophyll a, and temperature in png format. The function creates a directory called `plots/` in the file path to store the png's. Code expects only 2 excel workbooks for one monitoring run. Note sonde data from EXO models reads depth from the VPos metric. All other sonde models use a variant of a depth metric.

Usage

canning_surfR(path, obac, onic, SHELL = FALSE)

Arguments

path

Character string filepath to location of sonde data xlsx workbooks.

obac

Character string ("green", "blue" or "red") indicating oxygenation plant status at site BAC.

onic

Character string ("green", "blue" or "red") indicating oxygenation plant status at site NIC.

SHELL

Logical (TRUE/FALSE) indicating whether to include the Shelley Bridge site.

Value

A png four panel surfer plot of the Canning River. Saved to a directory called `plots/`.

Details

Surfer plots display a cross-section of the river where the metrics of interest have been interpolated between sonde locations. Thin plate spline has been used for the interpolation.

A river "bottom" is displayed which puts the interpolation in context. The river bottom has been derived from a combination of historical maximum depths at sampling locations as well as station points extracted from a "best" navigation line over the latest corporate bathymetry. See project documentation for further details.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/rivRmon/index.html the rivRmon website

Examples

if (FALSE) { # \dontrun{
canning_surfR(path = "Z:/DEC/MonitoringProgram/Data", obac = "green", onic = "red", SHELL = TRUE)
} # }