Creates summary HAB data for conditionally formatted table export

hab_groupR takes a file path to raw phyto data xlsx spreadsheets and creates summary data for use HAB reporting tables.

hab_groupR(pathin, pathout, skip = 5, mngt_triggers, SHELL = FALSE)

Arguments

pathin

a character filepath to the location of the raw phytoplankton xlsx spreadsheets.

pathout

a character filepath to the location of the desired directory for the summary csv files, suggest same as `pathin`

mngt_triggers

a character filepath and name of the `mngt_response_triggers_2020.csv` data file that indicates group names, species names and trigger thresholds for management action. See package example dataset `mngt_response_triggers` for example of format required.

SHELL

logical. Include Canning River site "Shell". Defaults to FALSE, change if required

Details

This is a precursor function to be run prior to using (hab_tablR). It takes a file path to raw xlsx phytoplankton data, creates cummulative counts on specific reportable phytoplankton groups and outputs a summary csv file. The groups are determined from an input file of management species and trigger levels.

Summary csv files are exported to a `HAB_tables/` directory and are named with the date, estuary and the words "HAB_data". If the directory doesn't exist, it is created.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/rivRmon/index.html the rivRmon website

Examples

if (FALSE) { # \dontrun{
phyto_groupR(pathin = "C:/path/to/raw_data", pathout = "C:/path/for_export",
mngt_triggers = "./mngt_response_triggers_2020.csv", skip = 5, SHELL = FALSE)
} # }