Creates summary LIMs data for interim reporting

lims_parsR takes a file path to raw LIMS xlsx spreadsheets and creates summary data.

lims_parsR(path, skip = 8)

Arguments

skip
numeric number of lines to skip at beginning of LIMS data ingest. Defaults to 8 which suits current LIMS format, change only if required.

pathin
a character filepath to the location of the week's raw LIMS data. Expects 3 CSV files, 2 for the Swan and one for the Canning. Data summary is written to this location.

Details

It takes a file path to 3 raw CSV files comprising of a week's worth of sampling as output from the Chem Centre. It then outputs a summary CSV data file, categorising chlorophyll a values into low, medium and high. "Low" is < 4ug, "medium" is 4 to 10ug and higher represents the "high" category. Summary output is combined for both rivers. Output has first day's date of sampling week and "interim_pigment_result" in the name.

Examples

if (FALSE) {
lims_parsR(pathin = "C:/path/to/raw_data", skip = 8)
}