Management Response Triggers

This data set is an example of the format required for creating a table for ingestion to the (hab_groupR) and (hab_tablR) functions. Values in this dataset determine which species are summarised and what values constitute differing levels of management action.

mngt_response_triggers

Format

a data.frame with 5 variables:

table_ord

numeric index of the desired order of reportable groups

table_name

character representing the reportable group name as it is intended for the output

spp_name

character representing the exact species name/s that compromise the reportable group

surveillance

numeric values below this are categorised as surveillance

notification

numeric values equal to or above are categorised as notification

Details

Although this is an example of the current triggers (as at time of writing) it is intended that this be used as a template to guide the creation of trigger data that would be saved out as a csv for use with the HAB functions. Note that values falling between surveillance and notification are classified as investigate.

Examples

mngt_response_triggers

#> # A tibble: 12 x 5
#>    table_ord table_name         spp_name               surveillance notification
#>        <dbl> <chr>              <chr>                         <dbl>        <dbl>
#>  1         1 Alex spp           alexandrium_spp                   5           10
#>  2         2 Dino spp           dinophysis_acuminata             23          230
#>  3         2 Dino spp           dinophysis_caudata               23          230
#>  4         2 Dino spp           dinophysis_acuta                 23          230
#>  5         2 Dino spp           dinophysis_fortii                23          230
#>  6         3 karenia spp        karenia_spp                      50          500
#>  7         4 karlo spp          karlodinium_spp               10000        30000
#>  8         5 Pn seriata group   pseudonitzschia_seria~           10          100
#>  9         6 Pn delicatissima ~ pseudonitzschia_delic~          100          500
#> 10         6 Pn delicatissima ~ pseudonitzschia_spp             100          500
#> 11         7 H akashiwo         heterosigma_akashiwo          20000        40000
#> 12         8 F japonica         fibrocapsa_japonica           20000        40000