Creates summary phytoplankton data for plotting

phyto_groupR takes a file path to raw phytoplankton xlsx spreadsheets and creates summary data for use in plotting.

Usage

phyto_groupR(pathin, pathout, skip = 5)

Arguments

pathin

a character filepath to the location of the raw phytoplankton xlsx spreadsheets i.e. the data from thephyoplankton lab.

pathout

a character filepath to the location of the desired directory for the summary csv files.

Details

This is a precursor function to be run prior to using (phyto_plotR). It takes a file path to raw xlsx phytoplankton data, creates summary counts on specific name groups and outputs a summary csv file. The groups are, Chlorophytes, Cyanophytes, Diatoms, Dinoflagellates, Cryptophyta and Other as a catch all.

Summary csv files are exported to a `summaries/` directory and are named with the date, river name and the word "summary". If the directory doesnt exist, it is created.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/rivRmon/index.html the rivRmon website

Examples

if (FALSE) { # \dontrun{
phyto_groupR(pathin = "C:/path/to/raw_data", pathout = "C:/path/for_export", skip = 5)
} # }